# federal-holidays-api/src/main/resources/federal-holidays-openapi.yml
openapi: 3.0.3
info:
  title: Federal Holidays Service API
  description: |
    REST API for managing USA and Canada federal holidays.
    Supports CRUD operations, search, filtering, and bulk file upload (CSV/JSON).
    
    **Authentication:** All endpoints require JWT Bearer token.
    **Required Headers:** X-Correlation-ID (UUID), Authorization (Bearer JWT).
  version: 1.0.0
  contact:
    name: RBC Development Team

servers:
  - url: http://localhost:8080/federal-holidays-api
    description: Local development server
  - url: https://dev-api.rbc.com/federal-holidays-api
    description: Development environment
  - url: https://api.rbc.com/federal-holidays-api
    description: Production environment

tags:
  - name: Holidays
    description: Federal holiday management operations (CRUD)
  - name: Search
    description: Search and filter holidays by country, year, or date range
  - name: Upload
    description: Bulk upload holidays from CSV or JSON files

security:
  - bearerAuth: []

paths:
  /api/v1/holidays:
    post:
      tags:
        - Holidays
      summary: Create a new federal holiday
      description: |
        Creates a new federal holiday with validation for unique constraint (one holiday per country per date).
        Country names are normalized (e.g., USA, US, UNITED STATES → USA).
      operationId: createHoliday
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HolidayRequest'
            examples:
              usaHoliday:
                summary: USA Independence Day
                value:
                  holidayName: "Independence Day"
                  holidayDate: "2026-07-04"
                  country: "USA"
                  isRecurring: true
                  description: "Celebrates the adoption of the Declaration of Independence"
              canadaHoliday:
                summary: Canada Day
                value:
                  holidayName: "Canada Day"
                  holidayDate: "2026-07-01"
                  country: "CANADA"
                  isRecurring: true
                  description: "National day of Canada"
      responses:
        '201':
          description: Holiday created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayResponse'
              examples:
                created:
                  value:
                    id: 1
                    holidayName: "Independence Day"
                    holidayDate: "2026-07-04"
                    country: "USA"
                    isRecurring: true
                    description: "Celebrates the adoption of the Declaration of Independence"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'

    get:
      tags:
        - Holidays
      summary: Get all federal holidays with pagination
      description: |
        Retrieves all federal holidays for USA and Canada with optional pagination.
        Default returns first 100 records. Use offset and limit for pagination.
      operationId: getAllHolidays
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - name: offset
          in: query
          required: false
          description: Starting position (0-based, default 0)
          schema:
            type: integer
            format: int32
            default: 0
            example: 0
        - name: limit
          in: query
          required: false
          description: Maximum records to return (default 100)
          schema:
            type: integer
            format: int32
            default: 100
            example: 100
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HolidayResponse'
              examples:
                holidays:
                  value:
                    - id: 1
                      holidayName: "Independence Day"
                      holidayDate: "2026-07-04"
                      country: "USA"
                      isRecurring: true
                      description: "Celebrates the adoption of the Declaration of Independence"
                    - id: 2
                      holidayName: "Canada Day"
                      holidayDate: "2026-07-01"
                      country: "CANADA"
                      isRecurring: true
                      description: "National day of Canada"
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/holidays/{id}:
    get:
      tags:
        - Holidays
      summary: Get holiday by ID
      description: Retrieves a single federal holiday by its unique identifier.
      operationId: getHolidayById
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - name: id
          in: path
          required: true
          description: Holiday ID
          schema:
            type: integer
            format: int64
            example: 1
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Holidays
      summary: Update an existing holiday
      description: Updates all fields of an existing holiday. All fields are replaced with new values from request.
      operationId: updateHoliday
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - name: id
          in: path
          required: true
          description: Holiday ID to update
          schema:
            type: integer
            format: int64
            example: 1
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HolidayRequest'
            examples:
              update:
                value:
                  holidayName: "Independence Day (Updated)"
                  holidayDate: "2026-07-04"
                  country: "USA"
                  isRecurring: true
                  description: "Updated description"
      responses:
        '200':
          description: Holiday updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags:
        - Holidays
      summary: Delete a holiday
      description: Permanently deletes a holiday by ID. This operation cannot be undone.
      operationId: deleteHoliday
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - name: id
          in: path
          required: true
          description: Holiday ID to delete
          schema:
            type: integer
            format: int64
            example: 1
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Holiday deleted successfully (No Content)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/holidays/search:
    get:
      tags:
        - Search
      summary: Search holidays with flexible criteria
      description: |
        Supports two search modes:
        1. **Date Range Mode:** Provide startDate + endDate to get holidays in date range
        2. **Filter Mode:** Provide country and/or year to filter holidays
        
        **Note:** Date range mode takes precedence if both modes are provided.
      operationId: searchHolidays
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - name: country
          in: query
          required: false
          description: Filter by country (USA, CANADA)
          schema:
            type: string
            enum: [USA, CANADA]
            example: USA
        - name: year
          in: query
          required: false
          description: Filter by year (YYYY format)
          schema:
            type: integer
            format: int32
            minimum: 1900
            maximum: 2100
            example: 2026
        - name: startDate
          in: query
          required: false
          description: Start date for range search (ISO 8601 format YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2026-01-01"
        - name: endDate
          in: query
          required: false
          description: End date for range search (ISO 8601 format YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2026-12-31"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HolidayResponse'
              examples:
                dateRange:
                  summary: Search by date range
                  value:
                    - id: 1
                      holidayName: "Independence Day"
                      holidayDate: "2026-07-04"
                      country: "USA"
                      isRecurring: true
                      description: "Celebrates the adoption of the Declaration of Independence"
                countryYear:
                  summary: Search by country and year
                  value:
                    - id: 2
                      holidayName: "Canada Day"
                      holidayDate: "2026-07-01"
                      country: "CANADA"
                      isRecurring: true
                      description: "National day of Canada"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/holidays/upload:
    post:
      tags:
        - Upload
      summary: Bulk upload holidays from file
      description: |
        Upload multiple holidays from a CSV or JSON file.
        - **Max file size:** 10MB
        - **Supported formats:** CSV, JSON
        - **Partial success:** Some records may fail validation while others succeed
        - **CSV format:** holidayName, holidayDate, country, isRecurring, description
        - **JSON format:** Array of holiday objects matching HolidayRequest schema
      operationId: uploadHolidays
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV or JSON file containing holidays (max 10MB)
            examples:
              csvUpload:
                summary: CSV file upload
                description: "Example CSV content:\nholidayName,holidayDate,country,isRecurring,description\nIndependence Day,2026-07-04,USA,true,US national holiday"
      responses:
        '200':
          description: File processed (may contain partial failures)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
              examples:
                fullSuccess:
                  summary: All records imported successfully
                  value:
                    totalRecords: 10
                    successCount: 10
                    errors: []
                partialSuccess:
                  summary: Some records failed validation
                  value:
                    totalRecords: 10
                    successCount: 8
                    errors:
                      - "Row 3: Duplicate holiday for USA on 2026-07-04"
                      - "Row 7: Invalid country: MEXICO"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File size exceeded (max 10MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-07-04T12:34:56.789Z"
                status: 413
                error: "Payload Too Large"
                code: "FILE_SIZE_EXCEEDED"
                message: "Maximum upload size exceeded"
                path: "/api/v1/holidays/upload"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication service

  parameters:
    CorrelationIdHeader:
      name: X-Correlation-ID
      in: header
      required: true
      description: UUID for request tracing and correlation across services
      schema:
        type: string
        format: uuid
        example: "123e4567-e89b-12d3-a456-426614174000"

  schemas:
    HolidayRequest:
      type: object
      required:
        - holidayName
        - holidayDate
        - country
      properties:
        holidayName:
          type: string
          minLength: 1
          description: Name of the federal holiday
          example: "Independence Day"
        holidayDate:
          type: string
          format: date
          description: Date of the holiday (ISO 8601 format YYYY-MM-DD)
          example: "2026-07-04"
        country:
          type: string
          minLength: 1
          description: Country code (normalized to USA or CANADA)
          enum: [USA, CANADA]
          example: "USA"
        isRecurring:
          type: boolean
          nullable: true
          description: Whether the holiday recurs annually on the same date
          example: true
        description:
          type: string
          nullable: true
          description: Optional description of the holiday
          example: "Celebrates the adoption of the Declaration of Independence"

    HolidayResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the holiday
          example: 1
        holidayName:
          type: string
          description: Name of the federal holiday
          example: "Independence Day"
        holidayDate:
          type: string
          format: date
          description: Date of the holiday (ISO 8601 format YYYY-MM-DD)
          example: "2026-07-04"
        country:
          type: string
          description: Country code (USA or CANADA)
          example: "USA"
        isRecurring:
          type: boolean
          nullable: true
          description: Whether the holiday recurs annually
          example: true
        description:
          type: string
          nullable: true
          description: Description of the holiday
          example: "Celebrates the adoption of the Declaration of Independence"

    FileUploadResponse:
      type: object
      properties:
        totalRecords:
          type: integer
          format: int32
          description: Total number of records in the uploaded file
          example: 10
        successCount:
          type: integer
          format: int32
          description: Number of records successfully imported
          example: 8
        errors:
          type: array
          items:
            type: string
          description: List of error messages for failed records (empty if all succeeded)
          example:
            - "Row 3: Duplicate holiday for USA on 2026-07-04"
            - "Row 7: Invalid country: MEXICO"
      description: |
        Response for bulk upload operation. 
        failureCount can be calculated as (totalRecords - successCount).

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred (RFC 3339)
          example: "2026-07-04T12:34:56.789Z"
        status:
          type: integer
          format: int32
          description: HTTP status code
          example: 400
        error:
          type: string
          description: HTTP error reason phrase
          example: "Bad Request"
        code:
          type: string
          description: Application-specific error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Detailed error message
          example: "Holiday name cannot be blank"
        path:
          type: string
          description: Request path that generated the error
          example: "/api/v1/holidays"

  responses:
    BadRequest:
      description: Invalid request parameters or validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationError:
              summary: Validation error
              value:
                timestamp: "2026-07-04T12:34:56.789Z"
                status: 400
                error: "Bad Request"
                code: "VALIDATION_ERROR"
                message: "Holiday name cannot be blank"
                path: "/api/v1/holidays"
            invalidCountry:
              summary: Invalid country
              value:
                timestamp: "2026-07-04T12:34:56.789Z"
                status: 400
                error: "Bad Request"
                code: "INVALID_COUNTRY"
                message: "Country 'MEXICO' is not supported. Supported countries: USA, CANADA"
                path: "/api/v1/holidays"

    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2026-07-04T12:34:56.789Z"
            status: 401
            error: "Unauthorized"
            code: "INVALID_TOKEN"
            message: "Invalid or expired JWT token"
            path: "/api/v1/holidays"

    NotFound:
      description: Holiday not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2026-07-04T12:34:56.789Z"
            status: 404
            error: "Not Found"
            code: "HOLIDAY_NOT_FOUND"
            message: "Holiday not found with id: 999"
            path: "/api/v1/holidays/999"

    Conflict:
      description: Duplicate holiday (same country and date)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2026-07-04T12:34:56.789Z"
            status: 409
            error: "Conflict"
            code: "DUPLICATE_HOLIDAY"
            message: "Holiday already exists for USA on 2026-07-04"
            path: "/api/v1/holidays"